lrcli splitup
